refactor(config): extract platform path helpers into internal/config/platform
Move OS-specific config path resolution out of flat platform*.go files and
into a dedicated sub-package internal/config/platform:

Files moved / renamed:
- platform.go         → platform/manager.go  (constants + userConfigFilePath)
- platform_darwin.go  → platform/darwin.go
- platform_unix.go    → platform/unix.go
- platform_windows.go → platform/windows.go
- platform_test.go    → platform/manager_test.go  (path-resolution tests only)

Cross-platform helpers (isRemoteURL, isGitHubShorthand, expandPaths) and
their tests are consolidated in utils.go / utils_test.go (package config).

config.UserConfigFilePath() is kept as a thin wrapper so all existing callers
(cmd/, telemetry) require no changes.

docs/config.md updated: module tree, Platform Sub-package section, rule 3
and rule 5 updated to reference the new package layout.
