+++ internal/config/platform/manager.go
@@ -1,10 +1,9 @@
-package config
+package platform
 
 import (
 	"log/slog"
 	"os"
 	"path/filepath"
-	"strings"
 )
 
 // ── Constants ─────────────────────────────────────────────────────────────────
@@ -35,9 +34,9 @@ const (
 
 // GlobalConfigFilePath and userConfigDir are defined in the OS-specific
 // build-constraint files:
-//   platform_windows.go  (build: windows)
-//   platform_darwin.go   (build: darwin)
-//   platform_unix.go    (build: !windows && !darwin)
+//   windows.go  (build: windows)
+//   darwin.go   (build: darwin)
+//   unix.go    (build: !windows && !darwin)
 
 // UserConfigFilePath is the exported accessor for the per-user config path.
 // It is used by callers outside this package (e.g. the init wizard and
@@ -58,64 +57,4 @@ func userConfigFilePath() string {
 		dir = filepath.Join(home, defaultConfigHome)
 	}
 	return filepath.Join(dir, configDirName, configFileName)
-}
-
-// ── Cross-platform path expansion ────────────────────────────────────────────
-
-// isRemoteURL reports whether s is a remote URL (http, https, git@, ssh).
-func isRemoteURL(s string) bool {
-	return strings.HasPrefix(s, "http://") ||
-		strings.HasPrefix(s, "https://") ||
-		strings.HasPrefix(s, "git@") ||
-		strings.HasPrefix(s, "ssh://")
-}
-
-// isGitHubShorthand reports whether s is a GitHub owner/repo shorthand
-// (exactly one forward-slash, no scheme, not a local path).
-func isGitHubShorthand(s string) bool {
-	if isRemoteURL(s) ||
-		strings.HasPrefix(s, "./") || strings.HasPrefix(s, `.\`) ||
-		strings.HasPrefix(s, "../") || strings.HasPrefix(s, `..\`) ||
-		strings.HasPrefix(s, "~/") || strings.HasPrefix(s, `~\`) ||
-		strings.HasPrefix(s, "/") || filepath.IsAbs(s) {
-		return false
-	}
-	return len(strings.Split(s, "/")) == 2
-}
-
-// expandPaths expands ~ and relative paths in c, while leaving remote URLs and
-// GitHub shorthands (owner/repo) untouched.
-func (c *Config) expandPaths(baseDir string) {
-	home, _ := os.UserHomeDir()
-
-	expandPath := func(p string) string {
-		// Accept both ~/ (POSIX) and ~\ (Windows) as home-relative prefixes.
-		if strings.HasPrefix(p, "~/") || strings.HasPrefix(p, `~\`) {
-			return filepath.Join(home, p[2:])
-		}
-		// filepath.IsAbs("/posix/path") returns false on Windows;
-		// handle POSIX-style absolute paths explicitly so cross-platform
-		// config files (e.g. written on Linux, used on Windows) are preserved.
-		if filepath.IsAbs(p) || strings.HasPrefix(p, "/") {
-			return p
-		}
-		return filepath.Join(baseDir, p)
-	}
-
-	expanded := make(map[string]ConfigRepo, len(c.Repositories))
-	for path, repo := range c.Repositories {
-		expanded[expandPath(path)] = repo
-	}
-	c.Repositories = expanded
-
-	for i := range c.Skills {
-		src := c.Skills[i].Source
-		if !isRemoteURL(src) && !isGitHubShorthand(src) {
-			c.Skills[i].Source = expandPath(src)
-		}
-	}
-
-	for i := range c.MCPs {
-		c.MCPs[i].Target = expandPath(c.MCPs[i].Target)
-	}
 }
