+++ internal/config/platform/unix.go
@@ -1,6 +1,6 @@
 //go:build !windows && !darwin
 
-package config
+package platform
 
 import "os"
